pyproject.toml
[tool.poetry]
name = "hansyo-ssr"
version = "0.1.0"
description = "Simple speaker recognition"
authors = ["Sota Hosoi <hosoi.sota.zw@tut.jp>"]
readme = "README.md"
packages = [
	{ include = "hansyo_ssr" }
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
torch = "^2.0.1"
torchaudio = "^2.0.2"
lightning = "^2.0.2"
hydra-core = "^1.3.2"
wandb = "^0.15.3"
torchviz = "^0.0.2"
torcheval = "^0.0.6"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pyproject-flake8 = "^6.0.0.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ["py37", "py38", "py39", "py310"]

[tool.flake8]
max-line-length = 119
extend-ignore = ["E203"]

[tool.isort]
profile = "black"

